services/agno-core/pyproject.toml
[project]
name = "agno-core"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.1",
  "agno>=0.1.0",
  "duckduckgo-search>=6.2.6",
]

[project.scripts]
serve = "agnocore.app:run"
